Second attempt at XBMC MySQL support
Added the ability to use a centralized MySQL XBMC library database. As
of the right now, if the connection fails, the user will not be alerted.
If the movie already exists in the library, the user won't be alerted
and the movie won't be added to the queue.  I have kept the ability to
check in either the SQLite local database or connect to a MySQL server
hosted library.  The setup can be found on the Extras tab, as I did not
feel it made sense for this to be under Notifications.